+++ src/main/kotlin/com/kuvaszuptime/kuvasz/repositories/LatencyLogRepository.kt
@@ -21,7 +21,7 @@ class LatencyLogRepository(private val dslContext: DSLContext) {
         private const val P99 = .99
     }
 
-    fun insertLatencyForMonitor(monitorId: Int, latency: Int) {
+    fun insertLatencyForMonitor(monitorId: Long, latency: Int) {
         dslContext.insertInto(LATENCY_LOG)
             .set(
                 LatencyLogRecord()
@@ -31,7 +31,7 @@ class LatencyLogRepository(private val dslContext: DSLContext) {
             .execute()
     }
 
-    fun fetchByMonitorId(monitorId: Int): List<LatencyLogRecord> = dslContext
+    fun fetchByMonitorId(monitorId: Long): List<LatencyLogRecord> = dslContext
         .selectFrom(LATENCY_LOG)
         .where(LATENCY_LOG.MONITOR_ID.eq(monitorId))
         .fetch()
@@ -41,14 +41,14 @@ class LatencyLogRepository(private val dslContext: DSLContext) {
         .where(LATENCY_LOG.CREATED_AT.lessThan(limit))
         .execute()
 
-    fun deleteAllByMonitorId(monitorId: Int) = dslContext
+    fun deleteAllByMonitorId(monitorId: Long) = dslContext
         .delete(LATENCY_LOG)
         .where(LATENCY_LOG.MONITOR_ID.eq(monitorId))
         .execute()
 
     // Well well, that's not so performant in case of a really huge dataset. Definitely something that should be
     // improved in the future.
-    fun getLatencyPercentiles(monitorId: Int? = null): List<PercentileResult> = dslContext
+    fun getLatencyPercentiles(monitorId: Long? = null): List<PercentileResult> = dslContext
         .with("percentiles").`as`(
             selectDistinct(
                 LATENCY_LOG.MONITOR_ID.`as`("monitor_id"),
@@ -71,7 +71,7 @@ class LatencyLogRepository(private val dslContext: DSLContext) {
         )
         .from(table("percentiles").`as`("p1"))
         .join(table("percentiles").`as`("p2"))
-        .on(field("p1.monitor_id", Int::class.java).eq(field("p2.monitor_id", Int::class.java)))
+        .on(field("p1.monitor_id", Long::class.java).eq(field("p2.monitor_id", Long::class.java)))
         .where(field("p1.percentile", Double::class.java).greaterOrEqual(P95))
         .and(field("p2.percentile", Double::class.java).greaterOrEqual(P99))
         .groupBy(field("p1.monitor_id"))
@@ -80,7 +80,7 @@ class LatencyLogRepository(private val dslContext: DSLContext) {
 
 @Introspected
 data class PercentileResult(
-    val monitorId: Int,
+    val monitorId: Long,
     val p95: Int?,
     val p99: Int?
 )
